+++ samples/SceneTransitionLayoutDemo/src/com/android/compose/animation/scene/demo/notification/DemoNotifications.kt
@@ -86,8 +86,8 @@ private object NotificationOnlyInShadePicker : StaticElementContentPicker {
     override fun contentDuringTransition(
         element: ElementKey,
         transition: TransitionState.Transition,
-        fromContentZIndex: Float,
-        toContentZIndex: Float,
+        fromContentZIndex: Long,
+        toContentZIndex: Long,
     ): ContentKey {
         return pickSingleContentIn(contents, transition, element)
     }
@@ -106,8 +106,8 @@ private object NotificationThatCanBeOnLockscreenPicker : StaticElementContentPic
     override fun contentDuringTransition(
         element: ElementKey,
         transition: TransitionState.Transition,
-        fromContentZIndex: Float,
-        toContentZIndex: Float,
+        fromContentZIndex: Long,
+        toContentZIndex: Long,
     ): ContentKey {
         fun handleLockscreenShadeTransition(
             lockscreen: ContentKey,
